[Issue #94]: Support Date and Time types. (#95)
We added Date and Time data types in pixels-core, and tested them from pixels-presto and pixels-load using orders table in TPCH.

For both types, they are represented as 4-byte integer and runlength+delta encoded.

For Date, the 4-byte integer is the number of days from epoch (1970-1-1 0:0:0 UTC). For Time, the 4-byte integer is the number of millis in the day.

##Changes:
* implement column reader/writer/vector for date and time.

* implement stats and Presto support for Date and Time types.

* test and fix Date and Time types support.

* test and fix Date/Time type integration in presto.

+++ pixels-core/src/main/java/io/pixelsdb/pixels/core/encoding/RunLenByteEncoder.java
@@ -26,6 +26,8 @@ import java.io.IOException;
  * A encoder for a sequence of bytes.
  * A control byte is written before each run with positive values 0 to 127 meaning 2 to 129 repetitions.
  * If the bytes is -1 to -128, 1 to 128 literal byte values follow.
+ *
+ * @author guodong
  */
 public class RunLenByteEncoder
         extends Encoder
@@ -60,12 +62,12 @@ public class RunLenByteEncoder
     }
 
     @Override
-    public byte[] encode(byte[] values, long offset, long length)
+    public byte[] encode(byte[] values, int offset, int length)
             throws IOException
     {
         for (int i = 0; i < length; i++)
         {
-            write(values[i + (int) offset]);
+            write(values[i + offset]);
         }
         flush();
         byte[] result = output.toByteArray();
